# openapi.org.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "opencode",
    "description": "opencode api",
    "version": "1.0.0"
  },
  "paths": {
...
    "/project/current": {
      "get": {
        "operationId": "project.current",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": {
              "type": "string"
            }
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "clientID",
            "required": true
          }
        ],
        "description": "Get the current project",
        "responses": {
          "200": {
            "description": "Current project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          }
        }
      }
    },
...
    "/session/{id}/message/{messageID}": {
      "get": {
        "operationId": "session.message",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Session ID"
          },
          {
            "in": "path",
            "name": "messageID",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Message ID"
          }
        ],
        "description": "Get a message from a session",
        "responses": {
          "200": {
            "description": "Message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "info": {
                      "$ref": "#/components/schemas/Message"
                    },
                    "parts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Part"
                      }
                    }
                  },
                  "required": [
                    "info",
                    "parts"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      }
    },
...
    "/event": {
      "get": {
        "operationId": "event.subscribe",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": {
              "type": "string"
            }
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "clientID",
            "required": true
          }
        ],
        "description": "Get events",
        "responses": {
          "200": {
            "description": "Event stream",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Event.lsp.updated": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "const": "lsp.updated"
          },
          "properties": {
            "type": "object",
            "properties": {}
          }
        },
        "required": [
          "type",
          "properties"
        ]
      },
...
      "UserMessage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionID": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "const": "user"
          },
          "time": {
            "type": "object",
            "properties": {
              "created": {
                "type": "number"
              }
            },
            "required": [
              "created"
            ]
          },
          "summary": {
            "type": "object",
            "properties": {
              "title": {
                "type": "string"
              },
              "body": {
                "type": "string"
              },
              "diffs": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FileDiff"
                }
              }
            },
            "required": [
              "diffs"
            ]
          },
          "agent": {
            "type": "string"
          },
          "model": {
            "type": "object",
            "properties": {
              "providerID": {
                "type": "string"
              },
              "modelID": {
                "type": "string"
              }
            },
            "required": [
              "providerID",
              "modelID"
            ]
          },
          "system": {
            "type": "string"
          },
          "tools": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {
              "type": "boolean"
            }
          }
        },
        "required": [
          "id",
          "sessionID",
          "role",
          "time",
          "agent",
          "model"
        ]
      },
...
      "AssistantMessage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionID": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "const": "assistant"
          },
          "time": {
            "type": "object",
            "properties": {
              "created": {
                "type": "number"
              },
              "completed": {
                "type": "number"
              }
            },
            "required": [
              "created"
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProviderAuthError"
              },
              {
                "$ref": "#/components/schemas/UnknownError"
              },
              {
                "$ref": "#/components/schemas/MessageOutputLengthError"
              },
              {
                "$ref": "#/components/schemas/MessageAbortedError"
              },
              {
                "$ref": "#/components/schemas/APIError"
              }
            ]
          },
          "parentID": {
            "type": "string"
          },
          "modelID": {
            "type": "string"
          },
          "providerID": {
            "type": "string"
          },
          "mode": {
            "type": "string"
          },
          "path": {
            "type": "object",
            "properties": {
              "cwd": {
                "type": "string"
              },
              "root": {
                "type": "string"
              }
            },
            "required": [
              "cwd",
              "root"
            ]
          },
          "summary": {
            "type": "boolean"
          },
          "cost": {
            "type": "number"
          },
          "tokens": {
            "type": "object",
            "properties": {
              "input": {
                "type": "number"
              },
              "output": {
                "type": "number"
              },
              "reasoning": {
                "type": "number"
              },
              "cache": {
                "type": "object",
                "properties": {
                  "read": {
                    "type": "number"
                  },
                  "write": {
                    "type": "number"
                  }
                },
                "required": [
                  "read",
                  "write"
                ]
              }
            },
            "required": [
              "input",
              "output",
              "reasoning",
              "cache"
            ]
          },
          "finish": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "sessionID",
          "role",
          "time",
          "parentID",
          "modelID",
          "providerID",
          "mode",
          "path",
          "cost",
          "tokens"
        ]
      },
      "Message": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/UserMessage"
          },
          {
            "$ref": "#/components/schemas/AssistantMessage"
          }
        ]
      },
...
      "TextPart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionID": {
            "type": "string"
          },
          "messageID": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "const": "text"
          },
          "text": {
            "type": "string"
          },
          "synthetic": {
            "type": "boolean"
          },
          "ignored": {
            "type": "boolean"
          },
          "time": {
            "type": "object",
            "properties": {
              "start": {
                "type": "number"
              },
              "end": {
                "type": "number"
              }
            },
            "required": [
              "start"
            ]
          },
          "metadata": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          }
        },
        "required": [
          "id",
          "sessionID",
          "messageID",
          "type",
          "text"
        ]
      },
      "ReasoningPart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionID": {
            "type": "string"
          },
          "messageID": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "const": "reasoning"
          },
          "text": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          },
          "time": {
            "type": "object",
            "properties": {
              "start": {
                "type": "number"
              },
              "end": {
                "type": "number"
              }
            },
            "required": [
              "start"
            ]
          }
        },
        "required": [
          "id",
          "sessionID",
          "messageID",
          "type",
          "text",
          "time"
        ]
      },
...
      "FilePart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionID": {
            "type": "string"
          },
          "messageID": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "const": "file"
          },
          "mime": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/FilePartSource"
          }
        },
        "required": [
          "id",
          "sessionID",
          "messageID",
          "type",
          "mime",
          "url"
        ]
      },
...
      "ToolPart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionID": {
            "type": "string"
          },
          "messageID": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "const": "tool"
          },
          "callID": {
            "type": "string"
          },
          "tool": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/ToolState"
          },
          "metadata": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          }
        },
        "required": [
          "id",
          "sessionID",
          "messageID",
          "type",
          "callID",
          "tool",
          "state"
        ]
      },
      "StepStartPart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionID": {
            "type": "string"
          },
          "messageID": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "const": "step-start"
          },
          "snapshot": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "sessionID",
          "messageID",
          "type"
        ]
      },
      "StepFinishPart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionID": {
            "type": "string"
          },
          "messageID": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "const": "step-finish"
          },
          "reason": {
            "type": "string"
          },
          "snapshot": {
            "type": "string"
          },
          "cost": {
            "type": "number"
          },
          "tokens": {
            "type": "object",
            "properties": {
              "input": {
                "type": "number"
              },
              "output": {
                "type": "number"
              },
              "reasoning": {
                "type": "number"
              },
              "cache": {
                "type": "object",
                "properties": {
                  "read": {
                    "type": "number"
                  },
                  "write": {
                    "type": "number"
                  }
                },
                "required": [
                  "read",
                  "write"
                ]
              }
            },
            "required": [
              "input",
              "output",
              "reasoning",
              "cache"
            ]
          }
        },
        "required": [
          "id",
          "sessionID",
          "messageID",
          "type",
          "reason",
          "cost",
          "tokens"
        ]
      },
      "SnapshotPart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionID": {
            "type": "string"
          },
          "messageID": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "const": "snapshot"
          },
          "snapshot": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "sessionID",
          "messageID",
          "type",
          "snapshot"
        ]
      },
      "PatchPart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionID": {
            "type": "string"
          },
          "messageID": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "const": "patch"
          },
          "hash": {
            "type": "string"
          },
          "files": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "sessionID",
          "messageID",
          "type",
          "hash",
          "files"
        ]
      },
      "AgentPart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionID": {
            "type": "string"
          },
          "messageID": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "const": "agent"
          },
          "name": {
            "type": "string"
          },
          "source": {
            "type": "object",
            "properties": {
              "value": {
                "type": "string"
              },
              "start": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "end": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              }
            },
            "required": [
              "value",
              "start",
              "end"
            ]
          }
        },
        "required": [
          "id",
          "sessionID",
          "messageID",
          "type",
          "name"
        ]
      },
      "RetryPart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionID": {
            "type": "string"
          },
          "messageID": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "const": "retry"
          },
          "attempt": {
            "type": "number"
          },
          "error": {
            "$ref": "#/components/schemas/APIError"
          },
          "time": {
            "type": "object",
            "properties": {
              "created": {
                "type": "number"
              }
            },
            "required": [
              "created"
            ]
          }
        },
        "required": [
          "id",
          "sessionID",
          "messageID",
          "type",
          "attempt",
          "error",
          "time"
        ]
      },
      "CompactionPart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionID": {
            "type": "string"
          },
          "messageID": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "const": "compaction"
          }
        },
        "required": [
          "id",
          "sessionID",
          "messageID",
          "type"
        ]
      },
      "Part": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/TextPart"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "sessionID": {
                "type": "string"
              },
              "messageID": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "const": "subtask"
              },
              "prompt": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "agent": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "sessionID",
              "messageID",
              "type",
              "prompt",
              "description",
              "agent"
            ]
          },
          {
            "$ref": "#/components/schemas/ReasoningPart"
          },
          {
            "$ref": "#/components/schemas/FilePart"
          },
          {
            "$ref": "#/components/schemas/ToolPart"
          },
          {
            "$ref": "#/components/schemas/StepStartPart"
          },
          {
            "$ref": "#/components/schemas/StepFinishPart"
          },
          {
            "$ref": "#/components/schemas/SnapshotPart"
          },
          {
            "$ref": "#/components/schemas/PatchPart"
          },
          {
            "$ref": "#/components/schemas/AgentPart"
          },
          {
            "$ref": "#/components/schemas/RetryPart"
          },
          {
            "$ref": "#/components/schemas/CompactionPart"
          }
        ]
      },
...
      "Event": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/Event.installation.updated"
          },
          {
            "$ref": "#/components/schemas/Event.installation.update-available"
          },
          {
            "$ref": "#/components/schemas/Event.lsp.client.diagnostics"
          },
          {
            "$ref": "#/components/schemas/Event.lsp.updated"
          },
          {
            "$ref": "#/components/schemas/Event.message.updated"
          },
          {
            "$ref": "#/components/schemas/Event.message.removed"
          },
          {
            "$ref": "#/components/schemas/Event.message.part.updated"
          },
          {
            "$ref": "#/components/schemas/Event.message.part.removed"
          },
          {
            "$ref": "#/components/schemas/Event.permission.updated"
          },
          {
            "$ref": "#/components/schemas/Event.permission.replied"
          },
          {
            "$ref": "#/components/schemas/Event.session.status"
          },
          {
            "$ref": "#/components/schemas/Event.session.idle"
          },
          {
            "$ref": "#/components/schemas/Event.session.compacted"
          },
          {
            "$ref": "#/components/schemas/Event.file.edited"
          },
          {
            "$ref": "#/components/schemas/Event.todo.updated"
          },
          {
            "$ref": "#/components/schemas/Event.workflow.started"
          },
          {
            "$ref": "#/components/schemas/Event.workflow.step.started"
          },
          {
            "$ref": "#/components/schemas/Event.workflow.step.completed"
          },
          {
            "$ref": "#/components/schemas/Event.workflow.step.failed"
          },
          {
            "$ref": "#/components/schemas/Event.workflow.paused"
          },
          {
            "$ref": "#/components/schemas/Event.workflow.resumed"
          },
          {
            "$ref": "#/components/schemas/Event.workflow.completed"
          },
          {
            "$ref": "#/components/schemas/Event.workflow.failed"
          },
          {
            "$ref": "#/components/schemas/Event.workflow.cancelled"
          },
          {
            "$ref": "#/components/schemas/Event.command.executed"
          },
          {
            "$ref": "#/components/schemas/Event.session.created"
          },
          {
            "$ref": "#/components/schemas/Event.session.updated"
          },
          {
            "$ref": "#/components/schemas/Event.session.deleted"
          },
          {
            "$ref": "#/components/schemas/Event.session.diff"
          },
          {
            "$ref": "#/components/schemas/Event.session.error"
          },
          {
            "$ref": "#/components/schemas/Event.client-tool.request"
          },
          {
            "$ref": "#/components/schemas/Event.client-tool.registered"
          },
          {
            "$ref": "#/components/schemas/Event.client-tool.unregistered"
          },
          {
            "$ref": "#/components/schemas/Event.client-tool.executing"
          },
          {
            "$ref": "#/components/schemas/Event.client-tool.completed"
          },
          {
            "$ref": "#/components/schemas/Event.client-tool.failed"
          },
          {
            "$ref": "#/components/schemas/Event.tui.prompt.append"
          },
          {
            "$ref": "#/components/schemas/Event.tui.command.execute"
          },
          {
            "$ref": "#/components/schemas/Event.tui.toast.show"
          },
          {
            "$ref": "#/components/schemas/Event.server.connected"
          },
          {
            "$ref": "#/components/schemas/Event.file.watcher.updated"
          }
        ]
      },
...
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "worktree": {
            "type": "string"
          },
          "vcs": {
            "type": "string",
            "const": "git"
          },
          "time": {
            "type": "object",
            "properties": {
              "created": {
                "type": "number"
              },
              "initialized": {
                "type": "number"
              }
            },
            "required": [
              "created"
            ]
          }
        },
        "required": [
          "id",
          "worktree",
          "time"
        ]
      },
...
      "BadRequestError": {
        "type": "object",
        "properties": {
          "data": {},
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "propertyNames": {
                "type": "string"
              },
              "additionalProperties": {}
            }
          },
          "success": {
            "type": "boolean",
            "const": false
          }
        },
        "required": [
          "data",
          "errors",
          "success"
        ]
      },
...
      "NotFoundError": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "const": "NotFoundError"
          },
          "data": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            },
            "required": [
              "message"
            ]
          }
        },
        "required": [
          "name",
          "data"
        ]
      },
...
    }
  }
}